Use regexp to match stdout instead of fmt.Sscanf
This allows us to match on goroutines conditionally and still support go
versions inferior to 1.4.

+++ tracedata.go
@@ -1,34 +1,34 @@
 package main
 
 type scvgtrace struct {
-	inuse    int
-	idle     int
-	sys      int
-	released int
-	consumed int
+	inuse    int64
+	idle     int64
+	sys      int64
+	released int64
+	consumed int64
 }
 
 type gctrace struct {
-	NumGC       int
-	Nproc       int
-	t1          int
-	t2          int
-	t3          int
-	t4          int
-	Heap0       int // heap size before, in megabytes
-	Heap1       int // heap size after, in megabytes
-	Obj         int
-	NMalloc     int
-	NFree       int
-	NSpan       int
-	NGoRoutines int
-	NBGSweep    int
-	NPauseSweep int
-	NHandoff    int
-	NHandoffCnt int
-	NSteal      int
-	NStealCnt   int
-	NProcYield  int
-	NOsYield    int
-	NSleep      int
+	NumGC       int64
+	Nproc       int64
+	t1          int64
+	t2          int64
+	t3          int64
+	t4          int64
+	Heap0       int64 // heap size before, in megabytes
+	Heap1       int64 // heap size after, in megabytes
+	Obj         int64
+	NMalloc     int64
+	NFree       int64
+	NSpan       int64
+	NGoRoutines int64
+	NBGSweep    int64
+	NPauseSweep int64
+	NHandoff    int64
+	NHandoffCnt int64
+	NSteal      int64
+	NStealCnt   int64
+	NProcYield  int64
+	NOsYield    int64
+	NSleep      int64
 }
